Updated the login and App.js
