+++ client/src/pages/ProfileUser.js
@@ -1,8 +1,12 @@
 import React from "react";
-import { Link } from "react-router-dom";
+import { Link, Redirect } from "react-router-dom";
 import { Typography, Button, Grid } from "@material-ui/core";
 
 function ProfileUser(props) {
+  if (!props.user) {
+    return <Redirect to='/' />
+  }
+  
   return (
     <div>
       <Grid container spacing={2}>
